feat: Process arbitrary datasets
